requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
gradio==4.12.0
langchain==0.0.350
chromadb==0.4.22
PyMuPDF==1.23.8
python-multipart==0.0.6
python-dotenv==1.0.0
ollama==0.1.6
langchain-community==0.0.10
python-docx==0.8.11
textract==1.6.5